Fixed up a bunch of things. Changed Properties to not have a bool anymore (toomuch headache with bool and const char * in a variant). Decided that having static holders of data is probably a bad idea especially if I have dll loaders/savers. They will not be able to access the applications data. Created the document class which will act as the data store house and also acts as the event generating class. So event changes are propigated to the Document class and then bubbled into the rest of the program.

+++ Model/Properties.PropertyTemplate.cpp
@@ -67,11 +67,11 @@ void Model::Properties::PropertyTemplate::bool_add(Key name, bool defaultvalue, 
 	auto f = find(name);
 	if (f != mDefs.end())
 	{
-		(*f).defvalue = defaultvalue;
+		(*f).defvalue = static_cast<long>(defaultvalue);
 		(*f).mandatory = mandatory;
 	}
 	else {
-		mDefs.emplace_back(name, mandatory, ValueTypes::vtBOOL, defaultvalue);
+		mDefs.emplace_back(name, mandatory, ValueTypes::vtBOOL, static_cast<long>(defaultvalue));
 	}
 }
